Calculate thermal time using cardinal temperatures

Calculate thermal time using cardinal temperatures

thermal_time(
  mint,
  maxt,
  x_temp = wheat$get("phenology.thermal_time.x"),
  y_temp = wheat$get("phenology.thermal_time.y"),
  method = "3hr"
)

Arguments

mint

The minimum temperature

maxt

The maximum temperature

x_temp

The cardinal temperatures using to calculate thermal time. The default values are c(0, 26, 37) defined in wheat$get("phenology.thermal_time.x").

y_temp

The effective thermal time at the corresponding cardinal temperatures. The default values are c(0, 26, 0) defined in wheat$get("phenology.thermal_time.y").

method

The method to calculate thermal time. The default method is "3hr", which uses the three hour temperature method. The other option is "default", which uses the default method ( maxt + mint ) / 2 - base.

Value

The thermal time.

Details

The daily thermal time is calculated by interpolating the effective thermal time at three hour intervals between the daily minimum and maximum temperatures, and then averaging these values to get the daily thermal time. For more information about the wheat phenology model, see the APSIM Next Generation Wheat Documentation.

Examples

mint <- c(0, 10)
maxt <- c(30, 40)
thermal_time(mint, maxt)
#> [1] 13.14622 11.68707